model_lab/video_classification/requirements.txt
av==8.0.3
certifi==2021.10.8
fvcore==0.1.5.post20211023
iopath==0.1.9
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
pandas==1.3.4
parameterized==0.8.1
Pillow==8.4.0
portalocker==2.3.2
python-dateutil==2.8.2
pytorchvideo==0.1.3
pytz==2021.3
PyYAML==6.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
tabulate==0.8.9
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
yacs==0.1.8
